FIX: add 'ko_Kore_KR.EUC-KR.php' to replace 'korean-euc-kr.php'.
I note the orifinal file is encoded by UTF-8 instead of it naming. I don't know why...

And '옯'  is out of the range of character set (KS X 1001) for EUC-KR. So this is replaced as '(out of range)'. In detail, please refer to RFC 1557 (http://datatracker.ietf.org/doc/rfc1557/).

git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1600 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b