FIX: i18n::$ref_table and two functions to keep compatibility
Till previous commit, different old language names match the same locale name. For example, 'english' and 'english-utf8' match 'en_Latn_US'. In English this is no matter but in the other languages such as 'japanese-utf8' and 'japanese-euc', this causes mis conversion with i18n::convert_locale_to_old_language_file_name(). 

To prevent from this, this commit identify 'language' string with hyphen-minus '-' as string with charset and preferentially match it with strings with charset.

This is reported in Japanese forum, special thanks to Reine!
http://japan.nucleuscms.org/bb/viewtopic.php?f=1&t=4935

git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1618 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b