FIX: replace Yen sign to backslash
Under working, we repeated the conversion between UTF-8 and SHIFT_JIS.
As a result, all of backslash is replaced Yen sign.
This is a commit to modify it.

Summary:
1. From UTF-8 to SHIFT_JIS, all of backslash character(0x5C) is recognized Yen sign.
2. From SHIFT_JIS to UTF-8, all yen sign is encoded into 0xC2A5.

As a result, there is no backslash and yen sign is a multi-byte character, not compatible for whole Character Encoding Scheme.


Detail:
USC-2, which is a character set of UTF-8, distinguish backslash and yen sign to assign different code points.
By UTF-8, the former is encoded to 0x5C and the latter is encoded to 0xC2A5.

JIS X 0201, which is one of character sets of SHIFT_JIS (the range of alphanumeric, punctuation and Hiragana), just define yen sign, don't define backslash.
By Shift_JIS, yen sign is encoded to 0x5c, as the same bynary for backslash in ASCII.

ISO 8859 series use the same characters for ASCII range.
So backslash is encoded to 0x5C.
But yen sign is defined in four character encoding schemes and then encoded to 0x5A. These are:
 ISO-8859-1
 ISO-8859-8
 ISO-8859-9
 ISO-8859-15

git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1707 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b