CHANGE: the timing to include Entity class
If sql_connect() fails to connect database, startupError() is called. But till 500 Internal Server Error is generated because Entity class is not included then.

FYI:
Since 4.0 release, Entity class becomes to be important class with some wrapper functions for htmlspechalchars/htmlentity PHP's built-in function. i18n and Entity class are good for multilingualization.

git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1711 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b