CHANGE: the available characters in table prefix, arranging message in translation files
In ANSI SQL89, the available characters in identifier is alphanumeric and underscore. But some hosting service in Japan give its usesr database named with hiphen-minus. SO we allowed hiphen-minus for $MYSQL_PREFIX and $MYSQL_DATABASE for 4,0 release.

But $MYSQL_PREFIX is surplus. In this commit, hiphen-minus in $MYSQL_PREFIX is unavailable.

The other modifications:
Some arrangements for messages in translation files.
Change class name from PARAM_MANAGER to ParamManager.


git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1713 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b