REMOVE: backtick (a.k.a backquote, grave accent) from identifier in SQL statement
For enabling to use under several SQL implementations such as SQLite, the sql statements generated by Nucleus CMS need to be compatible for several SQL implements.

For this issue, we select ANSI SQL89 standard. According the standard, identifier such as database name, table name and so on should consist of alphanumeric and underscore with no quotations. ANSI SQL92 just allow us to use reserved woords (a.k.a keywords) for identifier with double quotation.

The identifier with backticks is MySQL extension. In detail, please refer to MySQL document.
http://dev.mysql.com/doc/refman/5.0/en/identifiers.html

git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1714 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b