CHANGE: New Handler for Database is available and old sql_xxx functions are kept just for compatibility
add:
/nucleus/libs/sql/DB.php
/nucleus/libs/sql/MYSQLPDO.php
/nucleus/libs/sql/sql.php

remove:
/nucleus/libs/mysql.php
/nucleus/libs/sql/mysql.php
/nucleus/libs/sql/pdo.php

DB class is for whole database access.

case 1: $MYSQL_HANDLER = array('mysql','mysql') or array('mysql','')
In this case, MYSQLPDO class is used as a backend for DB class.
MYSQLPDO uses mysql_xxx() (or mysqli_xxx()) functions so this case is good for using old plugins which uses mysql_xxx() functions.

case 2: $MYSQL_HANDLER = array('pdo','mysql') or array('pdo','sqlite')
In this case, PHP's PDO implement is used as a backend for DB class.
So we can select the other handlers than mysql, for example, sqlite handler.

In both cases, sql_xxx() functions are served by /nucleus/libs/sql/sql.php, which is a wrapper for DB class. But this is not reccomended.

Plugins should use methods in DB class instead of sql_xxx() functions and of course must not use mysql_xxx() (or mysqli_xxx()) for portability.

git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1808 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b