CHANGE: adding NP_Medium as an alternative of media.php
media.php gives functions to deal with media files, uploading it and insert it into item.
But this is hard-coded and is not customizeable.
Additionally there are some alternative plugins and they are functionally integrated.

The aim of this commit is moving media functions into plugins and easy to create plugins.

git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1873 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b