ADD: customizable admin area (admin area skins)
Nucleus CMS's admin area is fixed by Admin class. But there are some demands to reconstruct for mobile device or text-read browser for the persons who feels inconvinience for their visibility.
http://sourceforge.net/tracker/?func=detail&aid=3523728&group_id=66479&atid=514646

Specification:
1. Utilizing Nucleus CMS's skin/template system for admin area.
2. admin area skin is re-distributable by the same way as normal skin
3. Batch/NavList/Encapsulate/PageFactory classes are needless because
they are implemented to AdminActions class and showlost.php
4. createaccount.php/forgotpassword.php is also skinnable.
5. media.php is deleted and replaced by NP_Medium plugin
6. All stuffs for admin area are in skin/admin directory
7. Fallback mechanism for failing to edit admin skin

This commits the changes for skin files for admin area

git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1888 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b