Improved - Default value when skin code value is empty
If each skin type code is empty, it is now used instead of the same skin
type code of the default blog. However, if there is only one blog site
or the skins of the default blog skin are empty, it becomes "skin
error".
In this case, we decided to use "index" type code of the same blog. In
this way, by using the <%if%> statement, depending on the configuration
of the site, it is possible to how to write skin code.

<html>
<head><title><%blogsetting(name)%></title></head>
<body>
<%if(skintype,index)%>
<%blog(default/index,10)%>
<%elseif(skintype,item)%>
<%item(default/item)%>
<%endif%>
</body>
</html>